… tests/proptests_lines_crlf.proptest-regressions