@cessen 2023-04-23
Thanks!  The general idea seems good (though I haven't reviewed the code closely).  I do want to keep the algorithm the same across architectures, however, just to keep the maintenance burden reasonable.  So if we can't find a performant way to implement `shift_across()` on x86_64, we might need to leave this on the table.