@CeleritasCelery 2023-04-24
I tried a basic approach to define it like this

```rust
    #[inline(always)]
    fn shift_across(&self, n: Self) -> Self {
        unsafe {
            let n = x86_64::_mm_slli_si128(n, 1);
            let upper = core::mem::transmute::<Self, [u8; Self::SIZE]>(*self);
            let mut lower = core::mem::transmute::<Self, [u8; Self::SIZE]>(n);
            lower[0] = upper[Self::SIZE - 1];
            core::mem::transmute::<[u8; Self::SIZE], Self>(lower)
        }
    }
```
but that resulted in a major 4x slowdown across all benchmarks. So it is going to need to be a different approach.